api/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
psycopg[binary]>=3.2.0
python-dotenv>=1.0.1
pytest>=8.2.0
httpx>=0.28.0
ruff>=0.6.0
